refactor(shared): split workTime, breaks and format into modules; index becomes pure re-export
- workTime.ts: gross/net/saldo calculations
- breaks.ts: manual breaks, legal minimum, applied break (German ArbZG sliding scale)
- format.ts: minutes/millis → TimeDuration

index.ts now contains only re-exports, so consumers don't need to know the
new file layout. Each new module fits in <60 lines and has a single concern.
